style: move 'reflection' folder into 'support'
