+++ cpp/persistent_stack.h
@@ -14,7 +14,7 @@
 
 #include "grammar_data_structure.h"
 #include "grammar_serializer.h"
-#include "reflection/reflection.h"
+#include "support/reflection.h"
 
 namespace xgrammar {
